backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
fastf1==3.4.0
anthropic==0.28.0
numpy==1.26.4
scipy==1.13.0
requests==2.32.0
pydantic==2.7.0
python-dotenv==1.0.1
